add memory efficient RNN predictions
Make lagged connections be resolved at runtime and use this feature
and a tiny bit of modulo arithmetic to reuse some of the first couple
of BPNs in CLUMPS over and over until the input sequence runs out.

Doesn't work for training because we need ot have everything around
for backpropagation.